server/requirements.txt
fastapi
uvicorn
python-dotenv
authlib
itsdangerous
jinja2
sqlalchemy
psycopg2-binary